fix: stats saving
