+++ src/components/SprintGameContent/SprintGameContent.tsx
@@ -1,13 +1,4 @@
 import { FC, useEffect } from 'react';
-import { useAppDispatch, useAppSelector } from '../../hooks/redux';
-import { GameStatsShort } from '../../interfaces/statistics';
-import {
-  useLazyGetStatisticsQuery,
-  useUpdateStatisticsMutation,
-} from '../../services/statisticsApi';
-import getCurrentDate from '../../shared/getCurrentDate';
-import { selectCurrentUserId } from '../../store/auth/authSlice';
-import { clearStats, selectStats, setStats } from '../../store/sprint/sprintSlice';
 import Button from '../Button';
 import s from './SprintGame.module.scss';
 
@@ -30,76 +21,6 @@ const SprintGameContent: FC<SprintGameProps> = ({
   checkWord,
   // isLoading,
 }) => {
-  const dispatch = useAppDispatch();
-
-  const [getStatistics] = useLazyGetStatisticsQuery();
-  const [updateStatistics] = useUpdateStatisticsMutation();
-
-  const userId = useAppSelector(selectCurrentUserId);
-  const stats = useAppSelector(selectStats);
-  const currentDate = getCurrentDate();
-
-  // SPRINT GAME STATS
-
-  useEffect(() => {
-    if (stats.date !== currentDate) {
-      dispatch(clearStats());
-    }
-
-    if (userId) {
-      const loadCurrentDateStats = async (): Promise<void> => {
-        const { data: currentStatsData } = await getStatistics(userId);
-
-        const currentDateStats =
-          currentStatsData?.optional?.games?.sprint?.filter((x) => x.date === currentDate)[0] ||
-          undefined;
-
-        if (currentDateStats) {
-          dispatch(setStats(currentDateStats));
-        }
-      };
-
-      loadCurrentDateStats();
-    }
-  }, []);
-
-  useEffect(() => {
-    if (seconds === 0 && userId) {
-      const updateStats = async (): Promise<void> => {
-        const { data: prevStatsData } = await getStatistics(userId);
-
-        const prevStats = prevStatsData?.optional?.games?.sprint || undefined;
-        const audioStats = prevStatsData?.optional?.games?.audiocall || undefined;
-        const currentDateStats = prevStats?.filter((x) => x.date === currentDate)[0] || undefined;
-
-        let sprint: GameStatsShort[];
-
-        if (prevStats) {
-          if (currentDateStats) {
-            sprint = [...prevStats.filter((x) => x.date !== currentDate), stats];
-          } else {
-            sprint = [...prevStats, stats];
-          }
-        } else {
-          sprint = [stats];
-        }
-
-        const body = {
-          optional: {
-            games: {
-              sprint,
-              audiocall: audioStats,
-            },
-          },
-        };
-
-        await updateStatistics({ userId, body });
-      };
-
-      updateStats();
-    }
-  }, [seconds]);
-
   // Action BTN
   useEffect(() => {
     const onKeypress = (e: KeyboardEvent): void => {
